% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{crossover}
\alias{crossover}
\title{Create crossover in a SNPbin object}
\usage{
crossover(snpbin)
}
\arguments{
\item{snpbin}{a SNPbin object}
}
\value{
a SNPbin object with its chromosomes crossed over
}
\description{
Create crossover in a SNPbin object
}
\details{
this will randomly choose a cut point in the chromosomes of a
  diploid snpbin object and create a crossover event.
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}

  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi:
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}
\keyword{internal}

