% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data}
\alias{example_data}
\title{A wrapper to get example data}
\usage{
example_data(set = "sample_data.feather")
}
\arguments{
\item{set}{the name of the data set you want to use. Defaults to "sample_data.feather"}
}
\value{
a string with the path to the example
}
\description{
A wrapper to get example data
}
\examples{
example_data()
}

