% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather2genind.R
\name{feather2genind}
\alias{feather2genind}
\title{Convert a feather file to a genind object}
\usage{
feather2genind(ff, locus_prefix = "Locus", sample = 50, genclone = TRUE,
  verbose = FALSE)
}
\arguments{
\item{ff}{a feather file on disk}

\item{locus_prefix}{An identifier for the locus columns (Default: "Locus")}

\item{sample}{the number of individuals to sample per population. (Default: 50)}

\item{genclone}{convert to a genclone object (Default: TRUE)}

\item{verbose}{report progress of import? (Default: FALSE)}
}
\value{
a genind object
}
\description{
Convert a feather file to a genind object
}
\examples{
library("tidyr")
ex  <- example_data()
gid <- feather2genind(ex)
# split the population name into its components
ex_run  <- "twenty_loci([0-9]+?)/"
ex_seed <- "seed_([0-9]+?)_"
ex_sex  <- "sex_([0-9.]+?)_"
ex_gen  <- "gen_([0-9]+?)_"
ex_rep  <- "rep_([0-9]+?).pop"
strata(gid) <- strata(gid) \%>\%
  extract(pop, c("run", "seed", "sexrate", "gen", "rep"),
          paste0(ex_run, ex_seed, ex_sex, ex_gen, ex_rep),
          remove = FALSE)
setPop(gid) <- ~run/rep/seed
gid
}

