% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather2genlight.R
\name{feather2genlight}
\alias{feather2genlight}
\title{Convert a feather file to a genind object}
\usage{
feather2genlight(ff, locus_regex = "^[0-9]+?_[0-9]+?$", sample = 50,
  snpclone = TRUE, verbose = FALSE)
}
\arguments{
\item{ff}{a feather file on disk}

\item{locus_regex}{a valid regex string denoting what a locus column looks like.}

\item{sample}{the number of individuals to sample per population. (Default: 50)}

\item{snpclone}{convert to a snpclone object (Default: TRUE)}

\item{verbose}{report progress of import? (Default: FALSE)}
}
\value{
a genlight object
}
\description{
Convert a feather file to a genind object
}
\examples{
\dontrun{
# Note: the examples don't exactly work at the moment :/
library("tidyr")
ex  <- example_data("sample_genomic_data.feather")
gid <- feather2genind(ex, sample = NULL)
# split the population name into its components
ex_run  <- "twenty_loci([0-9]+?)/"
ex_seed <- "seed_([0-9]+?)_"
ex_sex  <- "sex_([0-9.]+?)_"
ex_gen  <- "gen_([0-9]+?)_"
ex_rep  <- "rep_([0-9]+?).pop"
strata(gid) <- strata(gid) \%>\%
  extract(pop, c("run", "seed", "sexrate", "gen", "rep"),
          paste0(ex_run, ex_seed, ex_sex, ex_gen, ex_rep),
          remove = FALSE)
setPop(gid) <- ~run/rep/seed
gid
}
}

