% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_maf.R
\name{filter_maf}
\alias{filter_maf}
\title{Filter loci that don't have the correct minor allele frequencies}
\usage{
filter_maf(x, maf = 0.05, e = .Machine$double.eps)
}
\arguments{
\item{x}{a genlight or snpclone object}

\item{maf}{the minor allele frequency (default to 0.05)}

\item{e}{an error term added to maf}
}
\value{
a genlight or snpclone object
}
\description{
Filter loci that don't have the correct minor allele frequencies
}
\examples{
ex <- structure(c(1L, 0L, 0L, 1L, 0L, 1L, 1L, 0L, 1L, 0L, 0L, 1L, 0L,
      0L, 0L, 0L, 0L, 0L, 0L, 0L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L,
      2L, 1L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 1L, 2L, 0L, 0L, 0L, 0L, 0L,
      0L, 0L, 0L, 0L, 0L, 1L, 0L, 0L, 1L, 0L, 1L, 1L, 0L, 0L, 0L, 2L,
      2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 1L, 2L, 1L, 2L, 2L, 1L, 2L,
      2L, 1L, 2L, 1L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 1L, 0L, 0L, 0L, 0L,
      0L, 0L, 0L, 1L, 0L, 0L, 0L), .Dim = c(10L, 10L), .Dimnames = list(
          c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10"), c("0_0",
          "0_1", "0_2", "0_3", "0_4", "0_5", "0_6", "0_7", "0_8", "0_9"
          )))
gl <- new("genlight", ex)
gl
filter_maf(gl)
}

