% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_genlight.R
\name{genomic_ia}
\alias{genomic_ia}
\title{A test for significance with genomic ia}
\usage{
genomic_ia(x, sample = 0, blocksize = 100, quiet = FALSE)
}
\arguments{
\item{x}{a genlight or snpclone object}

\item{sample}{the number of samples for significance testing}

\item{blocksize}{the size of the blocks at which genotypes should be shuffled.}
}
\value{
a list with the observed value and pvalue and the samples used.
}
\description{
A test for significance with genomic ia
}
\note{
This is a slow procedure. When tested on 185 samples over 10,000 SNPs,
this took about 5 seconds/sample, which is 84 minutes of computing time on
my mac.
}
\examples{
gl <- glSim(100, 200, 0, parallel = FALSE)
genomic_ia(gl, sample = 999)
}

