% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{getNames}
\alias{getNames}
\title{Generate random sample names}
\usage{
getNames(n)
}
\arguments{
\item{n}{the number of random names to generate}
}
\description{
Generates random sample names by sampling 10 letters with replacement.
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary 
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for 
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}
  
  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of 
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi: 
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}

