% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather2genind.R
\name{ia_permute}
\alias{ia_permute}
\title{permute the index of assocation and return a tibble}
\usage{
ia_permute(gid, ...)
}
\arguments{
\item{gid}{a genind or genclone object}

\item{...}{any parameters to be passed to \code{\link[poppr]{ia}}}
}
\value{
a tibble from the ia object
}
\description{
permute the index of assocation and return a tibble
}
\examples{
library('poppr')
data(nancycats)
nancycats \%>\%
  seppop() \%>\%
  lapply(ia_permute, sample = 99, hist = FALSE) \%>\%
  bind_rows
}

