% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{mate}
\alias{mate}
\title{mate two individuals from a genlight object}
\usage{
mate(snpbin, ind1, ind2)
}
\arguments{
\item{snpbin}{a genlight object (yes, I know its mis-named).}

\item{ind1}{the index of parent 1}

\item{ind2}{the index of parent 2}
}
\value{
a SNPbin object
}
\description{
mate two individuals from a genlight object
}
\details{
this will take two parental samples (they do not have to be
  different), have them each undergo crossover, and then randomly sample one
  chromosome from each to create the offspring.
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}

  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi:
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}
\keyword{internal}

