% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{NA_generator}
\alias{NA_generator}
\title{Generate missing data for a SNPbin object}
\usage{
NA_generator(snpbin, nloc, na.perc = 0.01, rawchars = 2^(0:7))
}
\arguments{
\item{snpbin}{a SNPbin object}

\item{nloc}{the number of loci in the SNPbin object}

\item{na.perc}{the percentage of missing data to propogate}

\item{rawchars}{a vector representing masking bits. Passed for convenience.}
}
\description{
internal function utilized by pop_NA
}
\details{
the number of missing loci is generated here by drawing from a
  poisson distribution.
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary 
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for 
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}
  
  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of 
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi: 
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}

