% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{NA_zeromancer}
\alias{NA_zeromancer}
\title{Generate missing data in a single chromosome}
\usage{
NA_zeromancer(chrom, NA.posi, rawchars = 2^(0:7))
}
\arguments{
\item{chrom}{a vector of type raw.}

\item{NA.posi}{the position of missing data across the chromosome. Note that
these positions refer to positions of individual bits, not the individual
raw elements.}

\item{rawchars}{a vector representing the masking bits. Passed for
convenience.}
}
\description{
internal function utilized by NA_generator
}
\details{
This is the trickiest part to generate missing data. Missing data in
  genlight objects are represented by an integer vector representing the
  missing position, but in the binary coding, all missing data are
  represented as "off" bits. When generating missing data, all bits in the
  missing positions must be flipped to the "off" position using an XOR gate.
  Given a single NA position, when challenged with an "on" bit in an OR gate,
  if the result is "on" the same as the input, then the bit must be flipped
  because the input was in the "on" position.
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary 
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for 
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}
  
  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of 
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi: 
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}

