% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{pop_NA}
\alias{pop_NA}
\title{Propogate missing data in a genlight object}
\usage{
pop_NA(glt, na.perc = 0.01, parallel = require("parallel"), n.cores = 2L)
}
\arguments{
\item{na.perc}{the percentage of missing data per sample}

\item{parallel}{a logical specifying whether or not parallel processing
should be utilized}

\item{n.cores}{the number of cores to use with parallel processing.}
}
\value{
a genlight object with missing data propogated.
}
\description{
Propogate missing data in a genlight object
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}

  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi:
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}

