% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{random_mate}
\alias{random_mate}
\title{randomly mate a genlight object once}
\usage{
random_mate(glt, err)
}
\arguments{
\item{glt}{a genlight object}

\item{err}{the number of mutations per mating event.}
}
\value{
a genlight object with the same number of individuals as glt
}
\description{
randomly mate a genlight object once
}
\details{
Samples are chosen with replacement from the population twice to
  create each set of parents. They are then mated and mutated. See getSims
  for details
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary 
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for 
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}
  
  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of 
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi: 
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}

