% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{random_mate_gen}
\alias{random_mate_gen}
\title{randomly mate a genlight object for n generations.}
\usage{
random_mate_gen(glt, err = 0.005, gen = 1)
}
\arguments{
\item{glt}{a genlight object}

\item{err}{the number of mutations per mating event}

\item{gen}{the number of generations to mate the same population}
}
\value{
a genlight object
}
\description{
randomly mate a genlight object for n generations.
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary 
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for 
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}
  
  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of 
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi: 
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}

