% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Receiver Operator Characteristic}
\usage{
roc(alpha = 0.05, df, compare = c("0.0000", "1.0000"), stat = "p.rD",
  augment = NULL, count.na = TRUE, group = c("sexrate", "run", "seed",
  "sample"), flip = NULL)
}
\arguments{
\item{alpha}{the level at which to compare}

\item{df}{a data frame with p-values to compare}

\item{compare}{a two-element character vector specifying the levels of sexual reproduction to be the true positive and false positive.}

\item{stat}{What statistic should be compared}

\item{count.na}{When \code{TRUE}, missing values of \code{stat} are counted as positive}

\item{group}{the columns of data to group by}

\item{flip}{a column of logicals used to flip a negative to a positive. This
is used for extra criteria where the stat may not be considered significant
based on alpha, but can be considered significant based on the presence of
another factor}
}
\value{
a data frame
}
\description{
Receiver Operator Characteristic
}
\examples{
\dontrun{
# single value
x <- roc(df = vals, group = c('sexrate'))

# more granular
y <- roc(df = vals, group = c('sexrate', 'sample', 'mutation_rate'))

# very granular
z <- roc(df = vals, group = c('sexrate', 'sample', 'mutation_rate', 'run', 'seed'))

# Comparing clonal to sexual
res <- map_df(seq(0, 1, by = 0.01), roc, df = vals,
              group = c('sexrate', 'sample', 'mutation_rate'))
ggplot(res, aes(x = `False Positive`, y = `True Positive`, color = sample)) +
  geom_line() +
  geom_abline(slope = 1, lty = 2) +
  facet_wrap(~mutation_rate)
}
}

