% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_genlight.R
\name{sample_bitwise_ia}
\alias{sample_bitwise_ia}
\title{Calculate index of association from reshuffled genlight object}
\usage{
sample_bitwise_ia(i, x)
}
\arguments{
\item{i}{an integer, only used as a placeholder for the apply functions}

\item{x}{a genlight or snpclone object}
}
\value{
a value of the index of association
}
\description{
Calculate index of association from reshuffled genlight object
}
\examples{
gl <- glSim(100, 100, 100, parallel = FALSE)
bitwise.ia(gl)
sample_bitwise_ia(1, gl)
}

