% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Example simulation data}
\format{a tibble containing genclone objects}
\usage{
data(sample_data)
}
\description{
This data set contains simulations of clonal populations
}

