% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_genlight.R
\name{shuffle_genlight}
\alias{shuffle_genlight}
\title{resample a genlight/genclone object and produce a}
\usage{
shuffle_genlight(x)
}
\arguments{
\item{x}{a snpclone or genlight object}
}
\value{
a snpclone or genlight object
}
\description{
resample a genlight/genclone object and produce a
}
\examples{
gl <- glSim(100, 100, 100, parallel = FALSE)
gl
bitwise.ia(gl)
gls <- shuffle_genlight(gl)
gls
bitwise.ia(gls)
}

