% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontiers_code.R
\name{snp_mutator}
\alias{snp_mutator}
\title{Mutate a single SNP.}
\usage{
snp_mutator(chrom, mutation = sample(2^(0:7), 1))
}
\arguments{
\item{chrom}{a vector of raw elements}

\item{mutation}{a single integer representing a bit mask for a random
mutation.}
}
\value{
a raw vector with one bit changed.
}
\description{
One random SNP in one random chromosome. Used by sample_mutator.
}
\details{
the XOR gate is utilized to flip a single bit.
}
\author{
Zhian N. Kamvar
}
\references{
ZN Kamvar, JC Brooks, and NJ Grünwald. 2015. Supplementary 
  Material for Frontiers Plant Genetics and Genomics 'Novel R tools for 
  analysis of genome-wide population genetic data with emphasis on
  clonality'. DOI:
  \href{http://dx.doi.org/10.5281/zenodo.17424}{10.5281/zenodo.17424}
  
  Kamvar ZN, Brooks JC and Grünwald NJ (2015) Novel R tools for analysis of 
  genome-wide population genetic data with emphasis on clonality. Front.
  Genet. 6:208. doi: 
  \href{http://dx.doi.org/10.3389/fgene.2015.00208}{10.3389/fgene.2015.00208}
}

