% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ia.R
\name{tidy_ia}
\alias{tidy_ia}
\title{Generate a tidy data frame of the index of assocation}
\usage{
tidy_ia(gid, ..., verbose = TRUE, keepdata = TRUE)
}
\arguments{
\item{gid}{a genind or genclone object}

\item{...}{any parameters to be passed to \code{\link[poppr]{ia}}}

\item{verbose}{when \code{TRUE}, a status message will be printed at the
beginning of the analysis. \code{FALSE} suppresses this message.}

\item{keepdata}{when \code{TRUE}, the data set is kept in a column called
"dataset"}
}
\value{
a tibble from the ia object
}
\description{
Generate a tidy data frame of the index of assocation
}
\examples{
library('poppr')
data(nancycats)
nancycats \%>\%
  seppop() \%>\%
  lapply(tidy_ia, sample = 99, hist = FALSE) \%>\%
  bind_rows
}

