% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ia.R
\name{tidy_jackia}
\alias{tidy_jackia}
\title{Generate tidy data frame of jackknife statistics of the index of association}
\usage{
tidy_jackia(gid, ..., verbose = TRUE, strata = NA)
}
\arguments{
\item{gid}{a genind or genclone object}

\item{...}{parameters to be passed on to \code{\link[poppr]{jack.ia}}}

\item{verbose}{when \code{TRUE}, a status message will be printed at the
beginning of the analysis. \code{FALSE} suppresses this message.}

\item{strata}{the strata at which to clone-correct. Defaults to entire data
set.}
}
\value{
a tibble
}
\description{
Generate tidy data frame of jackknife statistics of the index of association
}
\examples{
library('poppr')
data(partial_clone)
strata(partial_clone) <- data.frame(pop = pop(partial_clone))
partial_clone \%>\%
  seppop() \%>\%
  lapply(tidy_jackia, reps = 999, strata = 1) \%>\%
  bind_rows()
}

