% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_locus_contribution.R
\name{tidy_locus_contribution}
\alias{tidy_locus_contribution}
\title{Gather statistics on locus contribution}
\usage{
tidy_locus_contribution(gid, n = 999, verbose = TRUE, ...)
}
\arguments{
\item{gid}{a genclone object}

\item{n}{the number of permutations for calculating the genotype accumulation curve.}

\item{verbose}{should this be noisy?}

\item{...}{arguments to be passed on to diversity_stats.}
}
\value{
a tibble object
}
\description{
Gather statistics on locus contribution
}
\examples{
data(monpop)
mll(monpop) <- "original"
monpop <- monpop \%>\%
  splitStrata(~Tree/Year/Symptom) \%>\%
  setPop(~Tree)
monpop \%>\%
  seppop() \%>\%
  lapply(tidy_locus_contribution) \%>\%
  bind_rows()
}

