% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_locus_table.R
\name{tidy_locus_table}
\alias{tidy_locus_table}
\title{Give a tidy locus table}
\usage{
tidy_locus_table(gid, ..., verbose = TRUE, cc = TRUE, strata = NA)
}
\arguments{
\item{gid}{a genclone object}

\item{...}{parameters to be passed on}

\item{verbose}{should this be noisy?}

\item{cc}{should the data be clone corrected as well?}

\item{strata}{The population strata to use with cc.}
}
\value{
a tibble
}
\description{
Give a tidy locus table
}
\examples{
data(nancycats)
  seppop(nancycats) \%>\%
  lapply(tidy_locus_table) \%>\%
  bind_rows()
}

