% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zksimanalysis-package.r
\docType{package}
\name{zksimanalysis}
\alias{zksimanalysis}
\alias{zksimanalysis-package}
\title{zksimanalysis.}
\description{
This package is for analysis of simulations created via simuPOP.
}
\section{functions}{


\itemize{
 \item \code{\link{feather2genind}}
 \item \code{\link{tidy_ia}}
 \item \code{\link{example_data}}
}
}

