% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather2genind.R
\name{example_data}
\alias{example_data}
\title{A wrapper to get example data}
\usage{
example_data()
}
\value{
a string with the path to the example
}
\description{
A wrapper to get example data
}
\examples{
example_data()
}

