% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listfun.R
\name{listfun}
\alias{listfun}
\title{Apply a function on a single element list}
\usage{
listfun(x, f, ...)
}
\arguments{
\item{x}{a single element list}

\item{f}{a function to apply to the first element of x}

\item{...}{any arguments that go with f}
}
\value{
the output of f
}
\description{
This allows one to work with list columns in dplyr
}
\examples{
x <- list(y = c(1:10, NA))
listfun(x, sum, na.rm = TRUE)
}

