% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Receiver Operator Characteristic}
\usage{
roc(alpha = 0.05, df, compare = c("0.0000", "1.0000"), stat = "p.rD",
  count.na = TRUE, group = c("sexrate", "run", "seed", "sample"))
}
\arguments{
\item{alpha}{the level at which to compare}

\item{df}{a data frame with p-values to compare}

\item{compare}{a two-element character vector specifying the levels of sexual reproduction to be the true positive and false positive.}

\item{stat}{What statistic should be compared}

\item{count.na}{When \code{TRUE}, missing values of \code{stat} are counted as positive}

\item{group}{the columns of data to group by}
}
\value{
a data frame
}
\description{
Receiver Operator Characteristic
}
\examples{
\dontrun{
lapply(seq(0, 1, 0.01), roc, vals, group = c("sexrate", "sample"), count.na = TRUE) \%>\%
  bind_rows \%>\%
  ungroup() \%>\%
  select(-sexrate) \%>\%
  spread(score, positive_fraction) -> y
ggplot(y, aes(x = Miss, y = Hit, color = sample)) +
  geom_line() +
  geom_point() +
  geom_abline(slope = 1, lty = 2)
}
}

