% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_genlight.R
\name{sample_bitwise_ia}
\alias{sample_bitwise_ia}
\title{Calculate index of association from reshuffled genlight object}
\usage{
sample_bitwise_ia(mat, x, blocksize = 100)
}
\arguments{
\item{mat}{a matrix derived from x}

\item{x}{a genlight or snpclone object}

\item{blocksize}{the size of the blocks at which genotypes should be shuffled.}
}
\value{
a value of the index of association
}
\description{
Calculate index of association from reshuffled genlight object
}
\examples{
gl <- glSim(100, 100, 100, parallel = FALSE)
bitwise.ia(gl)
sample_bitwise_ia(gl, as.matrix(gl))
}

