% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather2genind.R
\name{tidy_ia}
\alias{tidy_ia}
\title{permute the index of assocation and return a tibble}
\usage{
tidy_ia(gid, ...)
}
\arguments{
\item{gid}{a genind or genclone object}

\item{...}{any parameters to be passed to \code{\link[poppr]{ia}}}
}
\value{
a tibble from the ia object
}
\description{
permute the index of assocation and return a tibble
}
\examples{
library('poppr')
data(nancycats)
nancycats \%>\%
  seppop() \%>\%
  lapply(tidy_ia, sample = 99, hist = FALSE) \%>\%
  bind_rows
}

