% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_locus_table.R
\name{tidy_locus_table}
\alias{tidy_locus_table}
\title{Give a tidy locus table}
\usage{
tidy_locus_table(gid, ..., verbose = TRUE)
}
\arguments{
\item{gid}{a genclone object}

\item{...}{parameters to be passed on}

\item{verbose}{shoudl this be noisy?}
}
\value{
a tibble
}
\description{
Give a tidy locus table
}
\examples{
data(nancycats)
  seppop(nancycats) \%>\%
  lapply(tidy_locus_table) \%>\%
  bind_rows()
}

